+++ quadratic-core/src/controller/dependencies.rs
@@ -1,12 +1,53 @@
-use super::*;
+use std;
+
 use std::collections::HashSet;
 
+use crate::grid::Grid;
 use crate::Pos;
 use crate::Rect;
 
+impl Grid {
+    /// Given `cell` and `dependencies` adds a new node to the graph.
+    /// Returns the old dependencies of the node.
+    pub fn set_dependencies(
+        &mut self,
+        cell: Pos,
+        dependencies: Option<Vec<Rect>>,
+    ) -> Option<Vec<Rect>> {
+        // make sure cell is not in dependencies
+        if let Some(dependencies) = &dependencies {
+            if dependencies.iter().any(|rect| rect.contains(cell)) {
+                panic!("cell cannot depend on itself");
+            }
+        }
+
+        // update graph and return old dependencies
+        match dependencies {
+            Some(areas) => self.dependencies_mut().insert(cell, areas),
+            None => self.dependencies_mut().remove(&cell),
+        }
+    }
+
+    /// Returns cells that _directly_ depend on `area`.
+    /// Does not continue to traverse the graph.
+    pub fn get_dependent_cells(&mut self, area: Rect) -> HashSet<Pos> {
+        let mut seen = HashSet::new();
+
+        for node in self.dependencies_mut().iter() {
+            for rect in node.1.iter() {
+                if rect.intersects(area) {
+                    seen.insert(*node.0);
+                }
+            }
+        }
+
+        seen
+    }
+}
+
 #[test]
 fn test_graph() {
-    let mut cdc = ComputationDependencyController::new();
+    let mut cdc = Grid::new();
 
     cdc.set_dependencies(
         Pos { x: 3, y: 3 },
