--- quadratic-core/src/dgraph/mod.rs
@@ -1,6 +1,0 @@
-mod controller;
-
-pub use controller::*;
-
-#[cfg(test)]
-mod tests;
