+++ quadratic-core/src/grid/mod.rs
@@ -1,3 +1,5 @@
+use std::collections::HashMap;
+
 use lexicon_fractional_index::key_between;
 use serde::{Deserialize, Serialize};
 #[cfg(feature = "js")]
@@ -27,13 +29,13 @@ pub use formatting::{
 pub use ids::*;
 pub use sheet::Sheet;
 
-use crate::{dgraph::ComputationDependencyController, CellValue, Pos, Rect, Value};
+use crate::{CellValue, Pos, Rect, Value};
 
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
 #[cfg_attr(feature = "js", wasm_bindgen)]
 pub struct Grid {
     sheets: Vec<Sheet>,
-    dependencies: ComputationDependencyController,
+    dependencies: HashMap<Pos, Vec<Rect>>,
 }
 impl Default for Grid {
     fn default() -> Self {
@@ -44,16 +46,16 @@ impl Grid {
     pub fn new() -> Self {
         let mut ret = Grid {
             sheets: vec![],
-            dependencies: ComputationDependencyController::new(),
+            dependencies: HashMap::new(),
         };
         ret.add_sheet(None).expect("error adding initial sheet");
         ret
     }
-    pub fn dependencies_mut(&mut self) -> &mut ComputationDependencyController {
-        &mut self.dependencies
-    }
     pub fn sheets(&self) -> &[Sheet] {
         &self.sheets
+    }
+    pub fn dependencies_mut(&mut self) -> &mut HashMap<Pos, Vec<Rect>> {
+        &mut self.dependencies
     }
     pub fn sheet_ids(&self) -> Vec<SheetId> {
         self.sheets.iter().map(|sheet| sheet.id).collect()
